kris-core: move package to ch.difty.kris
