--- subprojects/kris-core/src/main/kotlin/com/gmail/gcolaianni5/jris/domain/RisType.kt
@@ -1,66 +1,0 @@
-package com.gmail.gcolaianni5.jris.domain
-
-/**
- * Types of References - as used in [RisTag] `TY`.
- *
- * The class was composed from information available on
- * [Wikipedia](https://en.wikipedia.org/wiki/RIS_(file_format)).
- */
-@Suppress("unused", "SpellCheckingInspection")
-enum class RisType(val description: String) {
-    ABST("Abstract"),
-    ADVS("Audiovisual material"),
-    AGGR("Aggregated Database"),
-    ANCIENT("Ancient Text"),
-    ART("Art Work"),
-    BILL("Bill"),
-    BLOG("Blog"),
-    BOOK("Whole book"),
-    CASE("Case"),
-    CHAP("Book chapter"),
-    CHART("Chart"),
-    CLSWK("Classical Work"),
-    COMP("Computer program"),
-    CONF("Conference proceeding"),
-    CPAPER("Conference paper"),
-    CTLG("Catalog"),
-    DATA("Data file"),
-    DBASE("Online Database"),
-    DICT("Dictionary"),
-    EBOOK("Electronic Book"),
-    ECHAP("Electronic Book Section"),
-    EDBOOK("Edited Book"),
-    EJOUR("Electronic Article"),
-    ELEC("Web Page"),
-    ENCYC("Encyclopedia"),
-    EQUA("Equation"),
-    FIGURE("Figure"),
-    GEN("Generic"),
-    GOVDOC("Government Document"),
-    GRANT("Grant"),
-    HEAR("Hearing"),
-    ICOMM("Internet Communication"),
-    INPR("In Press"),
-    JFULL("Journal (full)"),
-    JOUR("Journal"),
-    LEGAL("Legal Rule or Regulation"),
-    MANSCPT("Manuscript"),
-    MAP("Map"),
-    MGZN("Magazine article"),
-    MPCT("Motion picture"),
-    MULTI("Online Multimedia"),
-    MUSIC("Music score"),
-    NEWS("Newspaper"),
-    PAMP("Pamphlet"),
-    PAT("Patent"),
-    PCOMM("Personal communication"),
-    RPRT("Report"),
-    SER("Serial publication"),
-    SLIDE("Slide"),
-    SOUND("Sound recording"),
-    STAND("Standard"),
-    STAT("Statute"),
-    THES("Thesis/Dissertation"),
-    UNPB("Unpublished work"),
-    VIDEO("Video recording")
-}
