--- subprojects/kris-core/src/test/kotlin/com/gmail/gcolaianni5/jris/domain/RisTagTest.kt
@@ -1,95 +1,0 @@
-package com.gmail.gcolaianni5.jris.domain
-
-import org.junit.jupiter.api.Test
-
-@Suppress("SpellCheckingInspection")
-internal class RisTagTest {
-
-    @Test
-    @Suppress("LongMethod")
-    fun description() {
-        RisTag.values().map { it.description }.containsAll(listOf(
-            "Type of reference",
-            "First Author",
-            "Secondary Author",
-            "Tertiary Author",
-            "Subsidiary Author",
-            "Abstract",
-            "Author Address",
-            "Accession Number",
-            "Author",
-            "Location in Archives",
-            "This field maps to T2 for all reference types except for Whole Book and Unpublished Work references.",
-            "Custom 1",
-            "Custom 2",
-            "Custom 3",
-            "Custom 4",
-            "Custom 5",
-            "Custom 6",
-            "Custom 7",
-            "Custom 8",
-            "Caption",
-            "Call Number",
-            "This field can contain alphanumeric characters.",
-            "Title of unpublished reference",
-            "Place Published",
-            "Date",
-            "Name of Database",
-            "DOI",
-            "Database Provider",
-            "Editor",
-            "End Page",
-            "Edition",
-            "Reference ID",
-            "Issue number",
-            "Periodical name: user abbreviation 1.",
-            "Alternate Title",
-            "Periodical name: standard abbreviation.",
-            "Journal/Periodical name: full format.",
-            "Journal/Periodical name: full format.",
-            "Keywords",
-            "Link to PDF.",
-            "Link to Full-text.",
-            "Related Records.",
-            "Image(s).",
-            "Language",
-            "Label",
-            "Website Link",
-            "Number",
-            "Miscellaneous 2.",
-            "Type of Work",
-            "Notes",
-            "Abstract.",
-            "Number of Volumes",
-            "Original Publication",
-            "Publisher",
-            "Publishing Place",
-            "Publication year (YYYY)",
-            "Reviewed Item",
-            "Research Notes",
-            "Reprint Edition",
-            "Section",
-            "ISBN/ISSN",
-            "Start Page",
-            "Short Title",
-            "Primary Title",
-            "Secondary Title (journal title",
-            "if applicable)",
-            "Tertiary Title",
-            "Translated Author",
-            "Title",
-            "Translated Title",
-            "User definable 1.",
-            "User definable 2.",
-            "User definable 3.",
-            "User definable 4.",
-            "User definable 5.",
-            "URL",
-            "Volume number",
-            "Published Standard number",
-            "Primary Date",
-            "Access Date",
-            "End of Reference")
-        )
-    }
-}
