+++ subprojects/kris-guide/src/test/kotlin/com/gmail/gcolaianni5/jris/guide/GuideExamples.kt
@@ -2,12 +2,12 @@
 
 package com.gmail.gcolaianni5.jris.guide
 
-import com.gmail.gcolaianni5.jris.domain.RisRecord
-import com.gmail.gcolaianni5.jris.domain.RisType
+import ch.difty.kris.domain.RisRecord
+import ch.difty.kris.domain.RisType
+import ch.difty.kris.toRisLines
+import ch.difty.kris.toRisRecords
 import com.gmail.gcolaianni5.jris.accept
 import com.gmail.gcolaianni5.jris.process
-import com.gmail.gcolaianni5.jris.toRisLines
-import com.gmail.gcolaianni5.jris.toRisRecords
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.Flow
