+++ subprojects/kris-io/src/integrationTest/java/com/gmail/gcolaianni5/jris/usage/JRisIOUsageImportTest.java
@@ -6,12 +6,13 @@ import java.io.*;
 import java.util.Collections;
 import java.util.List;
 
+import ch.difty.kris.domain.RisRecord;
+import ch.difty.kris.domain.RisType;
 import com.gmail.gcolaianni5.jris.JRisIO;
-import com.gmail.gcolaianni5.jris.domain.RisRecord;
-import com.gmail.gcolaianni5.jris.domain.RisType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+@SuppressWarnings("SpellCheckingInspection")
 class JRisIOUsageImportTest {
 
     private File file;
