+++ subprojects/kris-io/src/integrationTest/kotlin/com/gmail/gcolaianni5/jris/JRisIOIntegrationTest.kt
@@ -1,7 +1,7 @@
 package com.gmail.gcolaianni5.jris
 
-import com.gmail.gcolaianni5.jris.domain.RisRecord
-import com.gmail.gcolaianni5.jris.domain.RisType
+import ch.difty.kris.domain.RisRecord
+import ch.difty.kris.domain.RisType
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.amshove.kluent.shouldContainAll
 import org.amshove.kluent.shouldHaveSize
