+++ subprojects/kris-io/src/integrationTest/kotlin/com/gmail/gcolaianni5/jris/usage/JRisIoUsageSpec.kt
@@ -1,9 +1,9 @@
 package com.gmail.gcolaianni5.jris.usage
 
+import ch.difty.kris.domain.RisRecord
+import ch.difty.kris.domain.RisType
 import com.gmail.gcolaianni5.jris.JRisIO
 import com.gmail.gcolaianni5.jris.accept
-import com.gmail.gcolaianni5.jris.domain.RisRecord
-import com.gmail.gcolaianni5.jris.domain.RisType
 import com.gmail.gcolaianni5.jris.process
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.InternalCoroutinesApi
